#1 поправил получение плагинов из ini файла
